endpoint: fix identity manager refCount issues
[ upstream commit f6a4eeb0cd9bf43ffecee3e260603c03805688a1 ]

This fixes a refCount issue in the identity manager in the case where an
agent is restoring endpoints that are in the process of being stopped
and cleaned up. In that case, the identity might never be added to the
identity manager, but it will unconditionally be removed from it when its
being cleaned up.

This is seemingly fine in the base case with a single
endpoint, but if there are two or more sharing the same identity, we are
approaching muddy water. In that case the identity will be
upserted to the identity manager by the endpoint that is running, and
then removed by the one that is not running. That results in the
selector policy being removed _and_ detached when the endpoint is
running. This has the potential to cause a policy deadlock on the
running endpoint, since it is running with a detached selector policy -
not getting new updates. Also, prior to
36c9bba ("policy/cache: simplify cache entry creation flow"), introduced
in v1.17, this also caused a permanently broken endpoint that is unable
to regenerate. In v1.17 and newer, the selector policy is instead leaked
when the endpoint then is deleted. This effectively caused that
mentioned patch to hide this bug, making it harder to notice.

We have seen this cause endpoints to be stuck for hours until they are
deleted, without being able to get new policy updates.

+++ pkg/endpoint/endpoint.go
@@ -262,6 +262,12 @@ type Endpoint struct {
 	// the endpoint's labels.
 	SecurityIdentity *identity.Identity `json:"SecLabel"`
 
+	// identitySet is used to keep track of whether the SecurityIdentity of an endpoint
+	// has been added to the identity manager. This is only false for new endpoints without an identity or
+	// on endpoints that have been restored. We do this to ensure we only remove it from
+	// the identity manager if we are sure it was added.
+	identitySet bool
+
 	// policyMapFactory is used to create endpoint policy maps
 	policyMapFactory policymap.Factory
 
@@ -1266,16 +1272,20 @@ func (e *Endpoint) leaveLocked(conf DeleteConfig) []error {
 	}
 
 	if !conf.NoIdentityRelease && e.SecurityIdentity != nil {
-		// Restored endpoint may be created with a reserved identity of 5
-		// (init), which is not registered in the identity manager and
-		// therefore doesn't need to be removed.
-		if e.SecurityIdentity.ID != identity.ReservedIdentityInit {
-			e.identityManager.Remove(e.SecurityIdentity)
-		}
+		if e.identitySet {
+			// Restored endpoint may be created with a reserved identity of 5
+			// (init), which is not registered in the identity manager and
+			// therefore doesn't need to be removed.
+			// If identity was never set, eg. if the endpoint was disconnecting when it was restored
+			// we have not yet added its identity - hence we should not remove it.
+			if e.SecurityIdentity.ID != identity.ReservedIdentityInit {
+				e.identityManager.Remove(e.SecurityIdentity)
+			}
 
-		_, err := e.allocator.Release(context.Background(), e.SecurityIdentity, false)
-		if err != nil {
-			errs = append(errs, fmt.Errorf("unable to release identity: %w", err))
+			_, err := e.allocator.Release(context.Background(), e.SecurityIdentity, false)
+			if err != nil {
+				errs = append(errs, fmt.Errorf("unable to release identity: %w", err))
+			}
 		}
 		e.removeNetworkPolicy()
 		e.SecurityIdentity = nil
